yt_fetcher/pyproject.toml
[tool.poetry]
name = "yt_fetcher"
version = "0.1.0"
description = ""
authors = ["legofrend <eremin.oleg@mail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.4.0"
python-json-logger = "^2.0.7"
sqlalchemy = "^2.0.33"
lxml = "^5.3.0"
aiohttp = "^3.10.5"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.10"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
google-api-python-client = "^2.138.0"
google-auth-oauthlib = "^1.2.1"


[tool.poetry.group.dev.dependencies]
pillow = "^11.0.0"
pandas = "^2.2.2"
jupyter = "^1.1.1"
beautifulsoup4 = "^4.12.3"
bs4 = "^0.0.2"
openai = "^1.47.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
